+++ src/java-apps/hawks/src/main/java/in/zycon/demo/hawks/models/News.java
@@ -1,0 +1,30 @@
+package in.zycon.demo.hawks.models;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class News {
+
+    Integer id;
+
+    String title;
+
+    String desc;
+
+    String imageUrl;
+
+    String url;
+
+    String pubDate;
+
+    public enum newsType {
+        SPORTS, WORLD, TECHNOLOGY
+    }
+
+    newsType newsType;
+}
