+++ src/java-apps/pom.xml
@@ -16,17 +16,21 @@
     <modules>
         <module>news</module>
         <module>gateway</module>
+        <module>hawks</module>
     </modules>
     <properties>
         <java.version>1.8</java.version>
     </properties>
     <build>
-        <plugins>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-            </plugin>
-        </plugins>
+        <pluginManagement>
+            <plugins>
+                <plugin>
+                    <groupId>org.springframework.boot</groupId>
+                    <artifactId>spring-boot-maven-plugin</artifactId>
+                </plugin>
+            </plugins>
+
+        </pluginManagement>
     </build>
 
 </project>
